same thing (#2)
* Create OLDREADME.md

* edit file to reflect everything and make some modifications to make it actually not put libraries in the wrong places

also for v0.0.0.0.3!
also the way I do version numbers is first number is when is after release* and when things old things break. second number is for several changes post release**. third is for post release** and for minor changes or when the package is finally out of the fourth number and is actually shippable without needing to have so much work but this may not apply for some projects where instead it can give you a small guide as well and tell you to code some way to do something or make you pick from a list that other people made. fourth is for pre-release*** and is for changes that are a bigger step in pre-release*** than the fifth number or is for a beta build and usually isn't used early on as every change early on in a project's life is going to probably break the previous version. and the fifth number is used for small changes in pre-release*** unless early on in the project's life where it is used as a way to signify the next package. and while you may think package mangers could be an issue, the answer is that for actual releases only the first 3 numbers are used and for any cases where the last two are used like beta builds or alpha builds will probably not fully work and really any pre-release*** (while beta builds will try to at least work, alpha builds are hit or miss and if the package works it will say so at the end of that package's information log)

*after release referring to after the project is no longer in pre-release for the first time which isn't used at first because I don't usually start on 1.0.0, and also most projects are good enough for usage when they are on a build on the fourth number and say they are functional (functional doesn't mean it doesn't expect you to code a method for it to work especially if the readme tells you "Coding skills are needed" it doesn't mean the package doesn't work but doesn't mean it does (the notice may be a bit different depending on the package and project and repo)), and the third number is the first time you may get a package manager version through a repo**** but it isn't always a thing that happens and no promises are made!

**post release refers to the same thing as after release but that one went on and on because of my adhd and need to say everything so here is the TL;DR: After the first time of pre-release*** is over for the project/repo****, but when "post release" takes affect is an issue with the third and fourth number.

***pre-release refers to either when a package is in beta/alpha for a upcoming next release while the project is after it's first stable build that I deem good to call no longer in pre-release for the project and now has a "release" that I can actually update and ship and work, or is the stage before the first "release" and when in that stage I'll use the fifth number for pushing builds as much as I can or when I want to push a build and commonly don't have any way to update without manually wiping the install locations and installing the new build (depending on project), and commonly doesn't function fully (while some packages don't function without you coding as well, and while that isn't common, the reason I keep bringing it up is because "usb-power-control" is one of those cases where I make a template for something instead of an application and in general it's easier to make templates, and while this project was originally meant to be for a school project and be an application with easy modification to suite your own materials, I tried very hard but I found that it's impossible to find a adapter for lighting that supports data through lightning at all let alone usb as well with otg so after spending 50au on a second adapter for lighting that I didn't even need but it advertised supporting data but it didn't and I kept it as I may have a use for it. so I'm moving usb-power control to a template instead of a application and I also will probably at one point make a project with a lot of templates and test implementations (all uses of "project" in this big amount of text refers to a github repository for a project, and for "****" every case of "repo" refers to repository for packages through a package manager) sorry for my adhd doing this, but I had to say it as I have been burning to say it the whole time while writing this!

**** (was mentioned in brackets in "***" which is just above, just look for '"****"' but without the single quotes but with the double quotes (like this "****") and just read that brackets section

sorry for my rambling!

Longest commit description ever!! and most of it isn't even related to the commit (some is but like....)

* Update README.md

* Update README.md

* part 1 of moving updates from README.md to not-part-of-program/updates.md and moving old/ to not-part-of-program/old/

* part 2 of moving updates from README.md to not-part-of-program/updates.md and moving old/ to not-part-of-program/old/

* Rename raspi_code_project_build_0.0.0.0.2.py to project_build_0.0.0.0.2.py

* Update raspi_code_project_build_0.2.0_ofd.py

* Final Part of the process

* update text

---------

Co-authored-by: Animmjfan <anthonyfalzon12@icloud.com>